Move DonationClient.kt out of ui package
